Fix:#2600 Change return of ponyint_hash_str from uint64_t to size_t
The use of uint64_t instead of size_t for ponyint_hash_str was causing
warning/erros on build.

Changed the return type to of ponyint_hash_str to unit64_t along with
its associated functions.

[Ticket: #2600]

+++ src/libponyc/pkg/package.c
@@ -116,7 +116,7 @@ static size_t package_hash(package_t* pkg)
   // Hash the full string instead of the stringtab pointer. We want a
   // deterministic hash in order to enable deterministic hashmap iteration,
   // which in turn enables deterministic package signatures.
-  return (size_t)ponyint_hash_str(pkg->qualified_name);
+  return ponyint_hash_str(pkg->qualified_name);
 }
